wip(#369): split giant TerminalCompositor block into lifecycle/resize/repaint/keypress; remove monolith
